Refactor: standardize `@Models` imports to lowercase `@models` for consistency ♻️
- Renamed `@Models` to `@models` across the codebase.
- Updated related imports, stubs, and service provider mappings to reflect the changes.
