+++ resources/js/inertia/models/Translatable.ts
