+++ resources/js/inertia/models/User.ts
