+++ src/Generators/Sources/ViewsGenerators/ReactTSPagesGenerator.php
@@ -147,7 +147,7 @@ class ReactTSPagesGenerator extends InertiaReactTSController
             ->componentName("Edit")
             ->formTitle("Edit {$this->table->modelNaming()}")
             ->componentProps("{{$this->table->variableNaming()}}:{{$this->table->variableNaming()}:{$this->table->modelNaming()}}")
-            ->import(new TsImportString($this->table->modelNaming(), "@/Models/{$this->table->modelNaming()}"))
+            ->import(new TsImportString($this->table->modelNaming(), "@/models/{$this->table->modelNaming()}"))
             ->action("post(route(\"{$updateRoute}\" , {$this->table->variableNaming()}.id));")
             ->when(
                 $this->table->hasTranslatableAttribute(),
